requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-environ==0.10.0
psycopg2-binary==2.9.6
drf-spectacular==0.26.2
pyTelegramBotAPI==4.11.0
